// src/app/module/user/user.service.ts
import status from "http-status";
import { Role, Speciality } from "../../../generated/prisma/client";
import AppError from "../../errorHelpers/AppError";
import { auth } from "../../lib/auth";
import { prisma } from "../../lib/prisma";
import { ICreateDoctorPayload } from "./user.interface";


const createDoctor = async (payload: ICreateDoctorPayload) =>{
    const specialities: Speciality[] =[] 
    for(const specialityID of payload.specialities){
        const speciality = await prisma.speciality.findUnique({
            where:{id:specialityID}
        })
        if(!speciality){
            // throw new Error(`Speciality with ID ${specialityID} not found`);
            throw new AppError(status.NOT_FOUND, `Speciality with ID ${specialityID} not found`);
        }
        specialities.push(speciality);  
    }

    const userExists = await prisma.user.findUnique({
        where:{email: payload.doctor.email}
    })

    if(userExists){
        // throw new Error('User with this email already exists');
        throw new AppError(status.CONFLICT, 'User with this email already exists');
    }

    const userData = await auth.api.signUpEmail({
        body:{
            email: payload.doctor.email,
            password: payload.password,
            role: Role.DOCTOR,
            name: payload.doctor.name,
            needPasswordChange: true
        }
    })

    try{
        const result = await prisma.$transaction(async(tx)=>{
            const doctorData = await tx.doctor.create({
                data:{
                    userId: userData.user.id,
                    ...payload.doctor
                     }
                      })
            const doctorSpecialtyData = specialities.map((speciality)=>{
                return {
                    doctorId: doctorData.id,
                    specialityId: speciality.id
                }
            })
            await tx.doctorSpeciality.createMany({
                data: doctorSpecialtyData
            })

            const doctor = await tx.doctor.findUnique({
                where:{id: doctorData.id},
                select:{
                    id: true,
                    name: true,
                    email: true,
                    profilePhoto: true,
                    contactNumber: true,
                    address: true,
                    registrationNumber: true,
                    experience: true,
                    gender: true,
                    appointmentFee: true,
                    qualification: true,
                    currentWorkingPlace: true,
                    designation: true,
                    createdAt: true,
                    updatedAt: true,
                    user:{
                        select:{
                            id: true,
                            email: true,
                            name: true,
                            role: true,
                            emailVerified: true,
                            isDeleted: true,
                            deletedAt: true,
                            createdAt: true,
                            updatedAt: true

                        }
                    },
                    specialities: {
                        select:{
                            speciality: {
                                select:{
                                    title: true,
                                    id: true
                                    
                                }
                            }
                        }
                    }
                }
            })

            return doctor;
        })

        return result;
    }catch(error){
        console.log("Transaction error: ", error);
        // Rollback - delete the created user
        await prisma.user.delete({
            where:{id: userData.user.id}
        });
        throw error;
    }
    

}


export const UserService = {
    createDoctor
}